backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openai-whisper==20240930
yt-dlp==2024.12.23
transformers==4.47.1
torch==2.5.1
accelerate==1.2.1
pydub==0.25.1
python-dotenv==1.0.1
